[WIN] Fix node module support
There are two issues here.

One is the ";" operator that does not exist in Windows. I chose "&&" to replace it since it functions the same on both systems. This has the side effect of not running the next command if the previous failed, but it looks like this behavior is fine.

The second is the use of repr on paths. This results in the Windows path separator being escaped, which is not accepted by the Windows shell. My solution was to remove the repr and place the string between double quotes to achieve the same effect, but gain compatibility.