Update README with PythonMonkey suggestion
This commit suggests the usage of PythonMonkey, a Python Library for executing JavaScript in Python. PythonMonkey's API is similar to JS2PYs and its JavaScript engine is fully compatible with the newest updates to JavaScript since it leverages SpiderMonkey. Another interesting side effect of it using SpiderMonkey under the hood is that it also supports WebAssembly for free.